feat: enable server-side rendering for SEO
Pages previously shipped an empty <div> to crawlers because SafeHydrate
skipped all rendering during SSR. Content pages now render full HTML on
the server and hydrate seamlessly on the client.

- Remove the SafeHydrate SSR gate; enable wagmi ssr mode and create the
  QueryClient per tree so server renders don't share cache state
- Server-fetch proposals (DB), proposal markdown (GitHub), validator
  groups and delegatees (RPC) and pass them to client components as
  React Query initialData, marked stale so clients refetch live values
  after hydration
- Serialize bigints across the server->client boundary with a marker
  encoding, since RSC prop serialization downgrades bigints to strings
  (and the BigInt.prototype.toJSON polyfill bypasses JSON replacers)
- Make browser-only hooks SSR-safe: useWindowSize, useIsMiniPay, and
  Identicon (jazzicon needs document); rework useTabs and useHistory to
  read the query string from window instead of useSearchParams, which
  would force a client-side rendering bailout on static pages
- Replace force-dynamic with revalidate=300 (ISR) on content pages and
  add richer metadata for governance, staking, and delegate pages
- Server data helpers return null on failure so pages gracefully fall
  back to client-side fetching
