+++ src/app/delegate/[address]/page.tsx
@@ -1,11 +1,14 @@
-'use server';
-
 // DO NOT USE "use client" here as it breaks metadata for openGraph
 import { Metadata } from 'next';
 import DelegatePage from 'src/features/delegation/components/delegatePage';
 import { getDelegateeMetadata } from 'src/features/delegation/delegateeMetadata';
+import { getServerSideDelegatees } from 'src/features/delegation/serverSideDelegatees';
 import { getXName } from 'src/features/delegation/utils';
 import { shortenAddress } from 'src/utils/addresses';
+import { serializeBigints } from 'src/utils/objects';
+
+// Serve a cached page and refresh it in the background at most every 5 minutes
+export const revalidate = 300;
 
 export type Params = Promise<{ address: Address }>;
 
@@ -13,15 +16,22 @@ export async function generateMetadata(props: { params: Params }): Promise<Metad
   const metadata = getDelegateeMetadata();
   const { address } = await props.params;
   const data = metadata[address];
+  const displayName = data?.name || shortenAddress(address);
+  const title = `${displayName} - ${shortenAddress(address)}`;
+  const description = data?.description
+    ? `Delegate to ${displayName} on Celo Mondo. ${data.description}`
+    : `Delegate to ${displayName} | ${address}`;
 
   return {
+    title,
+    description,
     openGraph: {
-      title: `${data.name} - ${shortenAddress(address)}`,
-      description: `Delegate to ${data.name} | ${address}`,
+      title,
+      description: `Delegate to ${displayName} | ${address}`,
     },
     twitter: {
       title: shortenAddress(address),
-      creator: getXName(data),
+      creator: data ? getXName(data) : undefined,
       site: '@celo',
       card: 'summary_large_image',
     },
@@ -30,5 +40,11 @@ export async function generateMetadata(props: { params: Params }): Promise<Metad
 
 export default async function Page(props: { params: Params }) {
   const { address } = await props.params;
-  return <DelegatePage address={address} />;
+  const initialDelegatees = await getServerSideDelegatees();
+  return (
+    <DelegatePage
+      address={address}
+      initialDelegatees={initialDelegatees ? serializeBigints(initialDelegatees) : undefined}
+    />
+  );
 }
