+++ src/app/layout.tsx
@@ -5,8 +5,6 @@ import { alpinaFont, interFont } from 'src/styles/fonts';
 import 'src/styles/globals.css';
 import { App } from './app';
 
-export const dynamic = 'force-dynamic';
-
 export const metadata: Metadata = {
   applicationName: 'Celo Mondo',
   metadataBase: new URL(links.home),
