+++ src/app/staking/[address]/page.tsx
@@ -1,21 +1,31 @@
-'use server';
-
 // DO NOT USE "use client" here as it breaks metadata for openGraph
 import { Metadata } from 'next';
 import ValidatorPage from 'src/features/staking/page';
+import { getServerSideValidatorGroups } from 'src/features/validators/serverSideValidatorGroups';
 import { shortenAddress } from 'src/utils/addresses';
+import { serializeBigints } from 'src/utils/objects';
+
+// Serve a cached page and refresh it in the background at most every 5 minutes
+export const revalidate = 300;
 
 type Params = Promise<{ address: Address }>;
 
 export async function generateMetadata(props: { params: Params }): Promise<Metadata> {
   const { address } = await props.params;
+  const groups = await getServerSideValidatorGroups();
+  const groupName = groups?.addressToGroup?.[address]?.name;
+  const displayName = groupName || shortenAddress(address);
+  const title = `Stake with ${displayName}`;
+  const description = `Stake CELO with ${displayName} (${address}) on Celo Mondo`;
   return {
+    title,
+    description,
     openGraph: {
-      title: `Celo Mondo | Stake with ${shortenAddress(address)}`,
-      description: `Celo Mondo | Stake with ${address}`,
+      title: `Celo Mondo | ${title}`,
+      description,
     },
     twitter: {
-      title: shortenAddress(address),
+      title: displayName,
       site: '@celo',
       card: 'summary_large_image',
     },
@@ -24,5 +34,11 @@ export async function generateMetadata(props: { params: Params }): Promise<Metad
 
 export default async function Page(props: { params: Params }) {
   const { address } = await props.params;
-  return <ValidatorPage address={address} />;
+  const initialData = await getServerSideValidatorGroups();
+  return (
+    <ValidatorPage
+      address={address}
+      initialValidatorGroups={initialData ? serializeBigints(initialData) : undefined}
+    />
+  );
 }
