+++ src/config/wagmi.tsx
@@ -11,7 +11,7 @@ import {
   walletConnectWallet,
 } from '@rainbow-me/rainbowkit/wallets';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { useMemo } from 'react';
+import { useMemo, useState } from 'react';
 import 'react-toastify/dist/ReactToastify.css';
 import { config, infuraRpcUrl } from 'src/config/config';
 import { Color } from 'src/styles/Color';
@@ -42,6 +42,7 @@ function buildWagmiConfig() {
     chains: [config.chain],
     connectors,
     syncConnectedChain: false, // only have 1 chain per deployment
+    ssr: true, // render a disconnected state on the server, reconnect after hydration
     transports: {
       [celo.id]: fallback([http(config.chain.rpcUrls.default.http[0]), http(infuraRpcUrl)]),
       [celoAlfajores.id]: http(config.chain.rpcUrls.default.http[0]),
@@ -49,10 +50,11 @@ function buildWagmiConfig() {
   });
 }
 
-const queryClient = new QueryClient();
-
 export function WagmiContext({ children }: { children: React.ReactNode }) {
   const wagmiConfig = useMemo(() => buildWagmiConfig(), []);
+  // One QueryClient per React tree; a module-level singleton would leak cached
+  // data between requests when rendering on the server
+  const [queryClient] = useState(() => new QueryClient());
 
   return (
     <WagmiProvider config={wagmiConfig}>
