--- src/utils/ssr.ts
@@ -1,10 +1,0 @@
-import { useEffect, useState } from 'react';
-
-// Is the component server-side rendering or not
-export function useIsSsr() {
-  const [isSsr, setIsSsr] = useState(true);
-  useEffect(() => {
-    setIsSsr(false);
-  }, []);
-  return isSsr;
-}
