+++ src/utils/useIsMiniPay.ts
@@ -1,7 +1,11 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 
 export function useIsMiniPay() {
-  // @ts-ignore
-  const [isMiniPay] = useState(() => typeof window !== 'undefined' && !!window.ethereum?.isMiniPay);
+  // Detected in an effect so server and first client render agree (no hydration mismatch)
+  const [isMiniPay, setIsMiniPay] = useState(false);
+  useEffect(() => {
+    // @ts-ignore window.ethereum is injected by the MiniPay wallet
+    setIsMiniPay(!!window.ethereum?.isMiniPay);
+  }, []);
   return isMiniPay;
 }
