+++ src/utils/useTabs.ts
@@ -1,19 +1,31 @@
-import { useRouter, useSearchParams } from 'next/navigation';
-import { useCallback } from 'react';
+'use client';
+import { useCallback, useEffect, useState } from 'react';
 
+// Reads/writes the ?tab= query param without useSearchParams so that pages
+// using tabs can still be statically rendered (useSearchParams would force a
+// client-side rendering bailout up to the nearest Suspense boundary).
 export default function useTabs<Tab extends string>(defaultTab: Tab) {
-  const router = useRouter();
-  const params = useSearchParams();
-  const tab = (params?.get('tab') as Tab) || defaultTab;
+  const [tab, setTab] = useState<Tab>(defaultTab);
 
-  const onTabChange = useCallback(
-    (nextTab: Tab) => {
-      if (params?.get('tab') !== nextTab) {
-        router.push('?tab=' + nextTab);
-      }
-    },
-    [router, params],
-  );
+  useEffect(() => {
+    const syncFromUrl = () => {
+      const param = new URLSearchParams(window.location.search).get('tab') as Tab | null;
+      setTab(param || defaultTab);
+    };
+    syncFromUrl();
+    window.addEventListener('popstate', syncFromUrl);
+    return () => window.removeEventListener('popstate', syncFromUrl);
+  }, [defaultTab]);
+
+  const onTabChange = useCallback((nextTab: Tab) => {
+    setTab(nextTab);
+    const url = new URL(window.location.href);
+    if (url.searchParams.get('tab') !== nextTab) {
+      url.searchParams.set('tab', nextTab);
+      // Shallow update: keep the URL shareable without re-running the route
+      window.history.pushState(null, '', url);
+    }
+  }, []);
 
   return {
     onTabChange,
